+++ src/channel/__init__.py
@@ -2,5 +2,5 @@ from .phase_modulator import PhaseModulator
 from .mzm import MZM
 from .optics import (coupler_split, coupler_combine, polarizer, polarization_rotator,
                      polarization_controller, pbs, beam_combiner, quarterwave, halfwave)
-from .fiber_sectional import (cable, apply_birefringence, apply_cd, apply_pmd,
-                               apply_attenuation, D_TOTAL, D_MATERIAL, D_WAVEGUIDE)
+from .fiber import (cable, apply_birefringence, apply_cd, apply_pmd,
+                    apply_attenuation, D_TOTAL, D_MATERIAL, D_WAVEGUIDE)